src/test/src/test/EnvTest.java
/*
package src.model;

import jason.asSyntax.Literal;
import jason.asSyntax.Structure;
import jason.asSyntax.parser.ParseException;
import jason.environment.grid.Location;
import org.junit.Before;
import org.junit.Test;

import java.util.logging.Level;
import java.util.logging.Logger;

import static org.junit.Assert.*;
import static org.mockito.Mockito.*;

public class EnvTest {
    private Env environment;
    private static final Logger log = Logger.getLogger(Env.class.getName());

    @Before
    public void setUp() {
        environment = new Env();
        environment.init(new String[]{"5"}); // Initialize with a sample argument
    }

    @Test
    public void testInitialization() {
        assertNotNull("Agent model should be initialized", environment.getAgentModel());
        assertNotNull("Agent view should be initialized", environment.getAgentView());
    }

    @Test
    public void testExecuteMoveAction() throws ParseException {
        Structure moveAction = Literal.parseLiteral("move");
        boolean result = environment.executeAction("agent0", moveAction);
        assertTrue("Move action should be handled", result);
    }

    @Test
    public void testExecuteGreetAction() throws ParseException {
        Structure greetAction = Literal.parseLiteral("greet('hello')");
        boolean result = environment.executeAction("agent0", greetAction);
        assertTrue("Greet action should be handled", result);
    }

    @Test
    public void testUnknownAction() throws ParseException {
        Structure unknownAction = Literal.parseLiteral("unknown");
        boolean result = environment.executeAction("agent0", unknownAction);
        assertFalse("Unknown action should not be handled", result);
    }

    @Test
    public void testHandleExceptions() {
        try {
            environment.init(null); // Pass invalid argument
            fail("Exception should have been thrown during initialization");
        } catch (Exception e) {
            log.log(Level.SEVERE, "Expected exception: " + e.getMessage(), e);
        }
    }
}
*/
